feat implement home page and aboutus
in this assignment, you will complete the SPA by implementing the home
and aboutus pages. You are provided with the starter files above that you
should download and put them into your app folder. At the end of this
 assignment you should have completed the following updates:

Implemented the necessary factory and functions in services.js
Implemented the controllers in controllers.js
Configured the corporate leaders information in aboutus.html
Configured the featured dish, promotions, and executive chef information in home.html