Assignment 4: Client-Server Communication
in this assignment, you will update the web application to get data from the server to render the information
 corresponding to the promotions and the leadership team of the company. In addition, you will handle the
 submission of the feedback form by posting the feedback to the server. At the end of this assignment,
 you should have completed the following:

Updated the promotion item in the home page to be rendered based on the data downloaded from the server.
Updated the home page and the aboutus page to render the information about the corporate
 leaders by downloading the data from the server
Enabled the users to submit feedback through the feedback form and recording their feedback on the server.